+++ plugin/src/mosaic/bregman/Tools.java
@@ -1949,7 +1949,7 @@ public class Tools
 		{
 			for (int i=0; i<ni; i++) {  
 				for (int j=0;j< nj; j++) {  	
-					temp[i][j]= (float) array[k][i][j];
+					temp[i][j] = array[k][i][j];
 				}
 			}
 			ImageProcessor imp= new FloatProcessor(temp);
